# backend/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Face2Une Project",
      "description": "API",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "/",
...
    "schemes": ["http", "https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/cert": {
          "get": {
            "tags": ["Certificates"],
            "summary": "Get all certificates in db",
            "parameters": [
              {
              "in": "query",
              "name": "saloon_id",
              "description": "id of studio(saloon) not related to the id you get in the response",
              "required": false,
              "type": "integer"              }
            ],
            "responses": {
              "200": {
                "description": "Returns all certificates in database in JSON format"
              }
            }
          }
        },
        "/cert/{certid}": {
          "get": {
            "tags": ["Certificates"],
            "summary": "Get a single certificate by id in db",
            "parameters": [
              {
              "in": "path",
              "name": "certid",
              "description": "id of certificate",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
            ],
            "responses": {
              "200": {
                "description": "Returns a single certificate in database in JSON format"
              }
            }
          }
        },
        "/create_cert": {
          "post": {
            "tags": ["Certificates"],
            "summary": "Create new certificate",
            "parameters": [
              {
              "in": "body",
              "name": "body",
              "description": "object of certificate",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Certificate"
              }
            }
            ],
            "responses": {
              "200": {
                "description": "Create a certificate in database using JSON format"
              }
            }
          }
        },
        "/update_cert": {
          "post": {
            "tags": ["Certificates"],
            "summary": "Update certificate by id",
            "parameters": [
              {
              "in": "body",
              "name": "body",
              "description": "object of certificate",
              "required": true,
              "schema": {
                "allOf": [
                  {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      }
                    }
                  },
                  {
                    "$ref": "#/definitions/Certificate"
                  }
                ]
              }
            }
            ],
            "responses": {
              "200": {
                "description": "Update a certificate in database using JSON format"
              }
            }
          }
        },
        "/delete_cert/{certid}": {
          "delete": {
            "tags": ["Certificates"],
            "summary": "Delete certificate by id",
            "parameters": [
              {
              "in": "path",
              "name": "certid",
              "description": "id of certificate",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
            ],
            "responses": {
              "200": {
                "description": "Delete a certificate in database by id"
              }
            }
          }
        },
        "/buy_cert": {
          "post": {
            "tags": ["Certificates"],
            "summary": "Buy certificate",
            "parameters": [
              {
              "in": "body",
              "name": "body",
              "description": "client id and certificate id",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "client_id": {
                    "type": "integer"
                  },
                  "certificate_id": {
                    "type": "integer"
                  },
                  "saloon_id": {
                    "type": "integer",
                    "default": 1
                  }
                }
              }
            }
            ],
            "responses": {
              "200": {
                "description": "Buy certificate for a client in database using JSON format"
              }
            }
          }
        },
...
        "/clients": {
          "get": {
            "tags": ["Clients"],
            "summary": "Get clients in db",
            "parameters": [
              {
                "in": "query",
                "name": "index",
                "description": "starting index(starting from 1)",
                "required": false,
                "default": 1,
                "schema": {
                  "type": "number"
                }
                
              },
              {
                "in": "query",
                "name": "per_page",
                "description": "max clients per page",
                "required": false,
                "default": 10,
                "schema": {
                  "type": "number"
                }
              },
              {
                "in": "query",
                "name": "filter_like",
                "description": "multifilter 'like' for all columns",
                "required": false,
                "default": "",
                "schema": {
                  "type": "string"
                }
              },
              {
                "in": "query",
                "name": "column",
                "description": "'equal' filter column name(id, full_name, phone, email, user_id))",
                "required": false,
                "default": "",
                "schema": {
                  "type": "string"
                }
              },
              {
                "in": "query",
                "name": "value",
                "description": "value of 'equal' filter column",
                "required": false,
                "default": "",
                "schema": {
                  "type": "string"
                }
              }
            ],
            "responses": {
              "200": {
                "description": "Returns clients in database in JSON format"
              }
            }
          }
        },
...
        "/clients/{clientid}": {
          "get": {
            "tags": ["Clients"],
            "summary": "Get a single client by id in db",
            "parameters": [
              {
              "in": "path",
              "name": "clientid",
              "description": "client id",
              "required": true,
              "schema": {
                "type": "integer"
              }
              }
            ],
            "responses": {
              "200": {
                "description": "Returns client in database in JSON format"
              }
            }
          }
        },
        "/create_client": {
          "post": {
            "tags": ["Clients"],
            "summary": "Create new client",
            "parameters": [
              {
              "in": "body",
              "name": "body",
              "description": "object of client",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Client"
              }
            }
            ],
            "responses": {
              "200": {
                "description": "Create a client in database using JSON format"
              }
            }
          }
        },
        "/update_client": {
          "post": {
            "tags": ["Clients"],
            "summary": "Update client by id",
            "parameters": [
              {
              "in": "body",
              "name": "body",
              "description": "object of client",
              "required": true,
              "schema": {
                "allOf": [
                  {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      }
                    }
                  },
                  {
                    "$ref": "#/definitions/Client"
                  }
                ]
              }
            }
            ],
            "responses": {
              "200": {
                "description": "Update a client in database using JSON format"
              }
            }
          }
        },
        "/delete_client/{clientid}": {
          "delete": {
            "tags": ["Clients"],
            "summary": "Delete client by id",
            "parameters": [
              {
              "in": "path",
              "name": "clientid",
              "description": "id of client",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
            ],
            "responses": {
              "200": {
                "description": "Delete a client in database by id"
              }
            }
          }
        },
        "/add_client_order": {
          "post": {
            "tags": ["Clients"],
            "summary": "Add order to client",
            "parameters": [
              {
              "in": "body",
              "name": "body",
              "description": "object of client order",
              "required": true,
              "schema": {
                "$ref": "#/definitions/ClientOrder"
              }
            }
            ],
            "responses": {
              "200": {
                "description": "Add order to client in database using JSON format"
              }
            }
          }
        },
...
         "/appoint": {
          "get": {
            "tags": ["Appointments"],
            "summary": "Get Appointments in db",
            "parameters": [
              {
              "in": "query",
              "name": "column",
              "description": "filter column name(id, procedure_id, reservation_date, client_id, total_price, reserved_on, saloon_id))",
              "required": false,
              "schema": {
                "type": "string"
              }
              },
              {
              "in": "query",
              "name": "value",
              "description": "value of filter column",
              "required": false,
              "schema": {
                "type": "string"
              }
              },
              {
                "in": "query",
                "name": "details",
                "description": "get details boolean",
                "required": false,
                "type": "boolean",
                "default": false
              }
            ],
            "responses": {
              "200": {
                "description": "Returns Appointments in database in JSON format"
              }
            }
          }
        },
        "/appoint/{appointid}": {
          "get": {
            "tags": ["Appointments"],
            "summary": "Get a single appointment by id in db",
            "parameters": [
              {
              "in": "path",
              "name": "appointid",
              "description": "id of appointment",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
            ],
            "responses": {
              "200": {
                "description": "Returns a single appointment in database in JSON format"
              }
            }
          }
        },
        "/create_appoint": {
          "post": {
            "tags": ["Appointments"],
            "summary": "Create new appointments",
            "parameters": [
              {
                "in": "body",
                "name": "body",
                "description": "object of appointments",
                "required": true,
                "schema":
                {
                  "type": "object",
                  "properties": 
                  {
                    "data": 
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/Appointment"
                      }
                    },
                    "promocode": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "promocode": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            ],
            "responses": {
              "200": {
                "description": "Returns session id for stripe payment"
              }
            }
          }
        },
        "/update_appoint": {
          "post": {
            "tags": ["Appointments"],
            "summary": "Update appointment by id",
            "parameters": [
              {
              "in": "body",
              "name": "body",
              "description": "object of appointment",
              "required": true,
              "schema": {
                "allOf": [
                  {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      }
                    }
                  },
                  {
                    "$ref": "#/definitions/Appointment"
                  }
                ]
              }
            }
            ],
            "responses": {
              "200": {
                "description": "Update an appointment in database using JSON format"
              }
            }
          }
        },
        "/delete_appoint/{appointid}": {
          "delete": {
            "tags": ["Appointments"],
            "summary": "Delete appointment by id",
            "parameters": [
              {
              "in": "path",
              "name": "appointid",
              "description": "id of appointment",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
            ],
            "responses": {
              "200": {
                "description": "Delete an appointment in database by id"
              }
            }
          }
        },
...
      },
    "definitions": {
...
      "Certificate": {
        "properties": {
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }, 
          "price_gbp": {
            "type": "number"
          }
        }
      },
...
      "Client": {
        "properties": {
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "required": false
          },
          "email": {
            "type": "string",
            "required": false
          },
          "user_id": {
            "type": "number",
            "default": null,
            "required": false
          }
        }
      },
      "ClientOrder": {
        "properties": {
          "client_id": {
            "type": "integer",
            "required": true
          },
          "email": {
            "type": "string"
          },
          "package_id": {
            "type": "integer",
            "default": null
          },
          "certificate_id": {
            "type": "number",
            "default": null
          },
          "saloon_id": {
            "type": "integer",
            "default": null
          }
        }
      },
      "Appointment": {
        "properties": {
          "procedure_id": {
            "type": "integer",
            "default": "1"
          },
          "additional_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "default": 7,
              "uniqueItems": true
            }
          },
          "reservation_date_time": {
            "type": "string",
            "format": "date-time"
          },
          "client_id": {
            "type": "integer"
          },
          "saloon_id": {
            "type": "int",
            "default": "1"
          }
        }
      },
...
    }
  }